Use the error handler with formatting of HTTP 422 validation errors
